refactor(user): migrating all business logic to service

+++ src/app/services/userService.js
@@ -1,23 +1,58 @@
+const ApiError = require("../../utils/ApiErrorUtils");
 const userRepository = require("../repositories/userRepository");
 
 function getAll(condition = {}) {
   return userRepository.getAll(condition);
 }
 
-const getOne = (condition = {}) => {
-  return userRepository.getOne(condition);
+const getOne = async (condition) => {
+  if (!condition) {
+    // bad request
+    throw new ApiError("getOne function need condition params", 400);
+  }
+
+  const user = await userRepository.getOne(condition);
+  if (!user) {
+    // resource not found
+    throw new ApiError("Can't find user's spesific data", 404);
+  }
+  return user;
 };
 
 const createUser = (newUser) => {
   return userRepository.createUser(newUser);
 };
 
-const deleteUser = (condition = {}) => {
-  return userRepository.deleteUser(condition);
+const softDeleteUser = async (condition) => {
+  const user = await getOne(condition);
+  if (!user) return;
+
+  // updating status to deleted and get deletedUser data
+  await updateUser({ status: "deleted" }, condition);
+
+  // soft-delete with paranoid
+  await userRepository.deleteUser(condition);
+
+  return getOne({ where: condition.where, paranoid: false });
 };
 
-const updateUser = (updatedData, condition = {}) => {
-  return userRepository.updateUser(updatedData, condition);
+const hardDeleteUser = async (condition) => {
+  const user = await getOne({ where: condition.where, paranoid: false });
+  if (!user) return;
+
+  await userRepository.deleteUser({ where: condition.where, force: true });
+
+  return user;
 };
 
-module.exports = { getAll, getOne, createUser, deleteUser, updateUser };
+const updateUser = async (updatedData, condition = {}) => {
+  const user = await getOne(condition);
+  if (!user) return;
+
+  // update user data
+  await userRepository.updateUser(updatedData, condition);
+
+  return getOne(condition);
+};
+
+module.exports = { getAll, getOne, createUser, softDeleteUser, hardDeleteUser, updateUser };
